# expression: implement vectorized evaluation for 'builtinIntervalIntSig'