<!--
Thank you for contributing to TiDB! Please read TiDB's [CONTRIBUTING](https://github.com/pingcap/community/blob/master/CONTRIBUTING.md) document **BEFORE** filing this PR.
-->

### What problem does this PR solve? <!--add issue link with summary if exists-->
Implement vectorized evaluation for builtinIntervalIntSig, for #12103

### What is changed and how it works?
```
BenchmarkVectorizedBuiltinCompareFunc/builtinIntervalIntSig-VecBuiltinFunc-8         	  100000	     23080 ns/op	       0 B/op	       0 allocs/op
BenchmarkVectorizedBuiltinCompareFunc/builtinIntervalIntSig-NonVecBuiltinFunc-8      	   30000	     44643 ns/op	       0 B/op	       0 allocs/op
BenchmarkVectorizedBuiltinCompareFunc/builtinIntervalIntSig-VecBuiltinFunc#01-8      	   50000	     29442 ns/op	       0 B/op	       0 allocs/op
BenchmarkVectorizedBuiltinCompareFunc/builtinIntervalIntSig-NonVecBuiltinFunc#01-8   	   30000	     49938 ns/op	       0 B/op	       0 allocs/op
BenchmarkVectorizedBuiltinCompareFunc/builtinIntervalIntSig-VecBuiltinFunc#02-8      	   50000	     38824 ns/op	       0 B/op	       0 allocs/op
BenchmarkVectorizedBuiltinCompareFunc/builtinIntervalIntSig-NonVecBuiltinFunc#02-8   	   30000	     59479 ns/op	       0 B/op	       0 allocs/op
```

### Check List <!--REMOVE the items that are not applicable-->

Tests <!-- At least one of them must be included. -->

 - Unit test